implement rate limiter directly in Fr3 and Panda struct
