+++ src/robot/fr3.rs
@@ -4,6 +4,7 @@ use crate::network::Network;
 use crate::robot::errors::FrankaErrors;
 use crate::robot::logger::Record;
 use crate::robot::private_robot::PrivateRobot;
+use crate::robot::rate_limiting::{RateLimiter, DELTA_T, LIMIT_EPS};
 use crate::robot::robot_data::{PrivateRobotData, RobotData};
 use crate::robot::robot_impl::RobotImplGeneric;
 use crate::robot::service_types;
@@ -18,13 +19,20 @@ use crate::robot::service_types::{
 use crate::robot::types::Fr3StateIntern;
 use crate::{Fr3Model, FrankaResult, RealtimeConfig, RobotState};
 use std::mem::size_of;
-use crate::robot::rate_limiting::Fr3RateLimiter;
 
+/// Maintains a network connection to the FR3 robot, provides the current robot state, gives access to
+/// the model library and allows to control the robot. See the [`Robot`](crate::Robot) trait for
+/// methods to control the robot.
+///
+/// # Rate Limiting
+/// For FR3 the rate limiter is turned off by default. You can enable it by setting the `limit_rate`
+/// argument in the control methods to `true`. The implementation is based on
+/// [https://frankaemika.github.io/docs/control_parameters.html#limits-for-franka-research-3](https://frankaemika.github.io/docs/control_parameters.html#limits-for-franka-research-3).
+/// That means the maximum velocity depends on the current joint configuration of the robot.
 pub struct Fr3(RobotImplGeneric<Self>);
 
 impl RobotData for Fr3 {
     type Model = Fr3Model;
-    type RateLimiterParameters = Fr3RateLimiter;
     type StateIntern = Fr3StateIntern;
     type State = RobotState;
 }
@@ -295,5 +303,171 @@ impl PrivateRobotData for Fr3 {
                 "libfranka-rs: Move command rejected due to activated safety function! Please disable all safety functions.",
             ))
         }
+    }
+}
+
+/// This is the position based velocity rate limiter for Fr3. The implementation is based on
+/// [https://frankaemika.github.io/docs/control_parameters.html#limits-for-franka-research-3](https://frankaemika.github.io/docs/control_parameters.html#limits-for-franka-research-3).
+impl Fr3 {
+    /// Tolerance value for joint velocity limits to deal with numerical errors and data losses.
+    pub const JOINT_VELOCITY_LIMITS_TOLERANCE: [f64; 7] = [
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[0],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[1],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[2],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[3],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[4],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[5],
+        LIMIT_EPS + Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_JOINT_ACCELERATION[6],
+    ];
+
+    fn upper_limits(a: f64, b: f64, c: f64, d: f64, e: f64, tolerance: f64) -> f64 {
+        f64::min(a, f64::max(0., b + f64::sqrt(f64::max(0., c * (d - e))))) - tolerance
+    }
+
+    fn lower_limits(a: f64, b: f64, c: f64, d: f64, e: f64, tolerance: f64) -> f64 {
+        f64::max(a, f64::min(0., b - f64::sqrt(f64::max(0., c * (d + e))))) + tolerance
+    }
+}
+
+impl RateLimiter for Fr3 {
+    const RATE_LIMITING_ON_PER_DEFAULT: bool = false;
+    const TOL_NUMBER_PACKETS_LOST: f64 = 0.;
+    const MAX_JOINT_JERK: [f64; 7] = [5000.0 - LIMIT_EPS; 7];
+    const MAX_JOINT_ACCELERATION: [f64; 7] = [10.0000 - LIMIT_EPS; 7];
+    const MAX_TRANSLATIONAL_JERK: f64 = 4500.0 - LIMIT_EPS;
+    const MAX_TRANSLATIONAL_ACCELERATION: f64 = 9.0000 - LIMIT_EPS;
+    const MAX_TRANSLATIONAL_VELOCITY: f64 = 3.0000
+        - LIMIT_EPS
+        - Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_TRANSLATIONAL_ACCELERATION;
+    const MAX_ROTATIONAL_JERK: f64 = 8500.0 - LIMIT_EPS;
+    const MAX_ROTATIONAL_ACCELERATION: f64 = 17.0000 - LIMIT_EPS;
+    const MAX_ROTATIONAL_VELOCITY: f64 = 2.5000
+        - LIMIT_EPS
+        - Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_ROTATIONAL_ACCELERATION;
+    const MAX_ELBOW_JERK: f64 = 5000. - LIMIT_EPS;
+    const MAX_ELBOW_ACCELERATION: f64 = 10.0000 - LIMIT_EPS;
+    const MAX_ELBOW_VELOCITY: f64 =
+        1.5 - LIMIT_EPS - Self::TOL_NUMBER_PACKETS_LOST * DELTA_T * Self::MAX_ELBOW_ACCELERATION;
+
+    fn compute_upper_limits_joint_velocity(q: &[f64; 7]) -> [f64; 7] {
+        [
+            Self::upper_limits(
+                2.62,
+                -0.3,
+                12.,
+                2.75010,
+                q[0],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[0],
+            ),
+            Self::upper_limits(
+                2.62,
+                -0.2,
+                5.17,
+                1.79180,
+                q[1],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[1],
+            ),
+            Self::upper_limits(
+                2.62,
+                -0.2,
+                7.,
+                2.90650,
+                q[2],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[2],
+            ),
+            Self::upper_limits(
+                2.62,
+                -0.3,
+                8.,
+                -0.1458,
+                q[3],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[3],
+            ),
+            Self::upper_limits(
+                5.26,
+                -0.35,
+                34.,
+                2.81010,
+                q[4],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[4],
+            ),
+            Self::upper_limits(
+                4.18,
+                -0.35,
+                11.,
+                4.52050,
+                q[5],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[5],
+            ),
+            Self::upper_limits(
+                5.26,
+                -0.35,
+                34.,
+                3.01960,
+                q[6],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[6],
+            ),
+        ]
+    }
+
+    fn compute_lower_limits_joint_velocity(q: &[f64; 7]) -> [f64; 7] {
+        [
+            Self::lower_limits(
+                -2.62,
+                -0.3,
+                12.,
+                2.75010,
+                q[0],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[0],
+            ),
+            Self::lower_limits(
+                -2.62,
+                -0.2,
+                5.17,
+                1.791800,
+                q[1],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[1],
+            ),
+            Self::lower_limits(
+                -2.62,
+                -0.2,
+                7.,
+                2.906500,
+                q[2],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[2],
+            ),
+            Self::lower_limits(
+                -2.62,
+                -0.3,
+                8.,
+                3.048100,
+                q[3],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[3],
+            ),
+            Self::lower_limits(
+                -5.26,
+                -0.35,
+                34.,
+                2.810100,
+                q[4],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[4],
+            ),
+            Self::lower_limits(
+                -4.18,
+                -0.35,
+                11.,
+                -0.54092,
+                q[5],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[5],
+            ),
+            Self::lower_limits(
+                -5.26,
+                -0.35,
+                34.,
+                3.019600,
+                q[6],
+                Self::JOINT_VELOCITY_LIMITS_TOLERANCE[6],
+            ),
+        ]
     }
 }
